% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CENoisyFit.R
\name{CENoisyFit}
\alias{CENoisyFit}
\title{Estimating a dynamic mixture via the noisy Cross-Entropy method}
\usage{
CENoisyFit(x, rawdata, rho, maxiter, alpha, nsim, nrepsInt, eps, r = 5)
}
\arguments{
\item{x}{list: sequence of integers 1,...,K, where K is the mumber of datasets. Set x = 1 in case
of a single dataset.}

\item{rawdata}{either a list of vectors or a vector: in the former case, each
vector contains a dataset to be used for estimation.}

\item{rho}{real in (0,1): parameter determining the quantile of the log-likelihood values to be used at each iteration.}

\item{maxiter}{non-negative integer: maximum number of iterations.}

\item{alpha}{real in (0,1): smoothing parameter.}

\item{nsim}{non-negative integer: number of replications used in the normal and lognormal updating.}

\item{nrepsInt}{non-negative integer: number of replications used in the Monte Carlo estimate of the normalizing constant.}

\item{eps}{non-negative real: tolerance for the stopping criterion of the noisy Cross-Entropy method.}

\item{r}{positive integer: length of window to be used in the stopping criterion.}
}
\value{
For each dataset, a list with the following elements is returned:

V (nreps x 12) matrix: updated mean and variance of the distributions used in the stochastic program.
nit (positive integer): number of iterations needed for convergence.
loglik (scalar): maximized log-likelihood.
}
\description{
This function estimates a dynamic mixture by means of the noisy Cross-Entropy method.
Currently only implemented for the lognormal - generalized Pareto case. This function
is mostly an auxiliary function, not suitable for the final user. Use CeNoisyFitBoot.R instead.
}
\details{
See Rubinstein and Kroese (2004, chap. 6).
}
\examples{
maxiter = 200
alpha = .5
rho = .05
eps = 1e-2
nsim = 1000
nrepsInt = 10000
res <- CENoisyFit(1,Metro2019,rho,maxiter,alpha,nsim,nrepsInt,eps)

}
\references{
{
\insertRef{rubio4}{FitDynMix}
}
}
\seealso{
CENoisyFitBoot
}
\keyword{cross-entropy.}
\keyword{dynamic}
\keyword{mixture;}
